% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agency_ports_withlatlong}
\alias{agency_ports_withlatlong}
\title{Data frame of agency port names and latitudes}
\format{
A data frame with 599 rows and 6 columns
\describe{
\item{agid_port}{Agency port abbreviation}
\item{agency_portdescription}{Agency port description}
\item{latitude}{Latitude of the port location with minutes and seconds in radians}
\item{longitude}{Longitude of the port location with minutes and seconds in radians}
\item{area}{North or south of 40.10 N. latitude}
\item{ifq_area}{IFQ area code used by the West Coast Groundfish Observer Program}
}
}
\usage{
agency_ports_withlatlong
}
\description{
Data frame of agency port names and latitudes
}
\keyword{datasets}
