% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch_for_expansions}
\alias{catch_for_expansions}
\title{Example table of catches by year, state, and geargroup}
\format{
A data frame with 36 rows and 4 columns
\describe{
\item{year}{A numeric string for landing year}
\item{state}{Character string for the landed state}
\item{geargroup}{A string denoting the gear type}
\item{catch_mt}{A numeric value for landings in metric tons}
}
}
\usage{
catch_for_expansions
}
\description{
Example table of catches by year, state, and geargroup
}
\keyword{datasets}
