% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGrades.R
\name{checkGrade}
\alias{checkGrade}
\title{Check Sample Numbers for PACFIN_GRADE_CODE}
\usage{
checkGrade(Pdata)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fisheries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}
}
\value{
todo: document what checkGrade returns.
}
\description{
Ensure that each sample number has only one PACFIN_GRADE_CODE of fish in it.
}
\author{
Kelli F. Johnson
}
