% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanColumns.R
\name{cleanColumns.bds}
\alias{cleanColumns.bds}
\title{Clean columns of PacFIN biological data}
\usage{
cleanColumns.bds(data)
}
\arguments{
\item{data}{A data frame with named columns.}
}
\value{
A data frame with fewer columns
}
\description{
Remove unnecessary columns from a data frame from PacFIN to reduce the
object size.
}
