% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPacFIN.R
\name{cleanPacFIN}
\alias{cleanPacFIN}
\title{Clean raw PacFIN data}
\usage{
cleanPacFIN(
  Pdata,
  keep_gears = NULL,
  keep_length_type = NULL,
  keep_sample_type = c("M"),
  keep_sample_method = "R",
  keep_age_method = NULL,
  keep_states = c("WA", "OR", "CA"),
  clean = TRUE,
  CLEAN = lifecycle::deprecated(),
  keep_INPFC = lifecycle::deprecated(),
  keep_missing_lengths = lifecycle::deprecated(),
  spp = NULL,
  verbose = TRUE,
  savedir = lifecycle::deprecated()
)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fisheries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{keep_gears}{A character vector including the gear types you want
to label as unique fleets. Order matters and will define fleet numbering.
If the argument is missing, which is the default, then all found gear
groups are maintained and ordered alphabetically. For more details see
\code{\link[=getGearGroup]{getGearGroup()}}, which lists a link where you can find the available gear
groupings and how they link to \code{"PACFIN_GEAR_CODE"} within your data. The vector
supplied to this argument should consist of only options available in
\code{unique(GearTable[["GROUP"]])}.
Typical entries will include character values
such as \code{HKL}, \code{POT}, \code{TWL}, where the latter is short for all non-shrimp
trawls and \code{TWS} is shrimp trawls. Other gear identification codes and
their definitions include \code{DRG} which is dredge gear, \code{MSC} which is all
other miscellaneous gear such as diving or river trawls, \code{NET} which is all
non-trawl net gear, \code{NTW} which is non-trawl gear, and \code{TLS} which is
trolling gear. As a special case, \code{MID} is available for spiny dogfish to
extract mid-water trawl data as a separate fleet.}

\item{keep_length_type}{A vector of character values specifying the types of
length samples to keep. There is no default value, though users will
typically want to keep \code{c("", "F", "A")}, but should also think about using
\code{c("", "F", "A", NA)}. Note that types other than those listed below can be
present, especially if you are dealing with a skate.
\code{A} is alternate length,
\code{D} is dorsal length,
\code{F} is fork length,
\code{S} is standard length, and
\code{T} is total length.}

\item{keep_sample_type}{A vector of character values specifying the types of
samples you want to keep. The default is to keep \code{c("M")}. Available types
include market (M), research (R), special request (S), and commercial
on-board (C). There are additional samples without a \code{SAMPLE_TYPE}, but
they are only kept if you include \code{NA} in your call. All sample types from
California are assigned to \code{M}. Including commercial on-board samples is
not recommended because they might also be in WCGOP data and would lead to
double counting.}

\item{keep_sample_method}{A vector of character values specifying the types
of sampling methods you want to keep. The default is to keep \code{"R"},
which refers to samples that were sampled randomly. Available types include
random (R), stratified (S), systematic (N), purposive (P), and special (X).}

\item{keep_age_method}{A vector of ageing methods to retain in the data. All
fish aged with methods other than those listed will no longer be considered
aged. A value of \code{NULL}, the default, will keep all ageing methods.
However, a vector of \verb{c("B", "BB", S", "", NA, 1, 2)} will keep all unaged
fish and those that were aged with break and burn and surface reads. You do
not really need to include such a verbose vector of values though because
numbers are converted to appropriate character codes in \code{\link[=getAge]{getAge()}}.
Therefore, something like \code{c("B", "S")} would be sufficient to keep all
break and burn and surface reads.}

\item{keep_states}{A vector of states that you want to keep, where each state
is defined using a two-letter abbreviation, e.g., \code{WA}. The default is to
keep data from all three states, \code{keep_states = c("WA", "OR", "CA")}. Add
\code{'UNK'} to the vector if you want to keep data not assigned to a state.}

\item{clean}{A logical value used when you want to remove data from the input
data set. The default is \code{TRUE}. Where the opposite returns the original
data with additional columns and reports on what would have been removed.}

\item{CLEAN}{\emph{Deprecated} A logical value used when you want to remove data from the input
data set. The default is \code{TRUE}. Where the opposite returns the original
data with additional columns and reports on what would have been removed.}

\item{keep_INPFC}{\emph{Deprecated}. Areas are now defined using different methods.}

\item{keep_missing_lengths}{\emph{Deprecated}. Just subset them using
\verb{is.na(Pdata[, 'length']) after running }cleanPacFIN` if you want to remove
lengths, though there is no need because the package accommodates keeping
them in.}

\item{spp}{A character string giving the species name to ensure that the
methods are species specific. Leave \code{NULL} if generic methods work for
your species. Currently, sablefish is the only species with
species-specific code.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}

\item{savedir}{Deprecated. This function no longer creates plots of cleaned data,
but you can call \code{\link[=plotCleaned]{plotCleaned()}} separately if you want to create these plots.}
}
\value{
The input data filtered for desired areas and record types
specified, with added columns
\itemize{
\item year: initialized from SAMPLE_YEAR\cr
\item fleet: initialized to 1
\item fishery: initialized to 1
\item state: initialized from AGENCY_CODE.  Change using \link{getState}
\item length: length in mm, where \code{NA} indicates length is not available
\item lengthcm: floored cm from FORK_LENGTH when available, otherwise FISH_LENGTH
\item geargroup: the gear group associated with each \href{http://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/gr.txt}{PACFIN_GEAR_CODE}
\item weightkg: fish weight in kg from FISH_WEIGHT and FISH_WEIGHT_UNITS
}
}
\description{
Clean raw PacFIN data to remove unsuitable samples if \code{clean = TRUE} and
convert units of measured quantities to work with downstream functions.
Raw data are meant to be inclusive of everything from PacFIN so users can
explore all that is available, but this means that raw data will \strong{ALWAYS}
include information that is not appropriate for use in
US West Coast stock assessments.
}
\details{
The original fields in the returned data are left untouched,
with the exception of
\itemize{
\item \code{SEX_CODE}: modified using \code{\link[nwfscSurvey:codify_sex]{nwfscSurvey::codify_sex()}} and upon return will
only include character values such that fish with an unidentified sex are
now \code{"U"}.
\item Age: the best ages to use going forward rather than just the first age read.
The data are put through various tests before they are returned
and the results of these tests are stored in the \code{clean} column.
Thus, sometimes it is informative to run \code{cleanPacFIN(clean = FALSE)}
and use frequency tables to inspect which groups of data will be removed
from the data set when you change the code to be \code{clean = TRUE}.
The \code{fleet}, \code{fishery}, and \code{state} columns are meant for use in
stratifying the data according to the particulars of an assessment.
}
}
\seealso{
\code{\link[=getState]{getState()}}, \code{\link[=getLength]{getLength()}}, \code{\link[=getAge]{getAge()}}, \code{\link[=getAgeMethod]{getAgeMethod()}}, \code{\link[=getweight]{getweight()}}, \code{\link[=getArea]{getArea()}}
}
\author{
Andi Stephens, Kelli F. Johnson, Chantel R. Wetzel
}
