% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertlength_skate.R
\name{convertlength_skate}
\alias{convertlength_skate}
\title{Convert Disc and Interspiracular Width to Length}
\usage{
convertlength_skate(Pdata, returntype = c("all", "estimated"))
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fisheries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{returntype}{A character value from the list supplied that specifies
the data you want returned:
"all" = all lengths with the new estimates replacing the input lengths
"estimated" = only the estimate lengths with all other lengths as NA such that the
returned vector is the same length and the number of rows of the input data frame.}
}
\value{
A vector of lengths.
}
\description{
Convert from disc and interspiracular width to length for skates,
Conversion parameters were derived from
West Coast Groundfish Bottom Trawl Survey data.
}
\author{
Kelli F. Johnson and Vladlena Gertseva
}
