% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EF1_Denominator.R
\name{EF1_Denominator}
\alias{EF1_Denominator}
\title{Calculate the denominator for the level-1 expansion factor.}
\usage{
EF1_Denominator(
  Pdata,
  fa,
  fb,
  ma,
  mb,
  ua,
  ub,
  verbose = TRUE,
  plot = lifecycle::deprecated(),
  col.weight = "weightkg",
  savedir = NULL
)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fisheries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{fa, ma, ua}{Female-, male-, and unsexed-specific weight--length
coefficients for Stock Synthesis where the relationships were calculated
using length in cm and weight in kg. There are no default values. You must
calculate these values and pass them to the function. If a particular sex
does not pertain to your data, then just pass \code{NA} for that relationship.}

\item{fb, mb, ub}{Female-, male, and unsexed-specific weight--length exponents
for Stock Synthesis where the relationships were calculated using length in
cm and weight in kg. If a particular sex does not pertain to your data,
then just pass \code{NA} for that relationship.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}

\item{plot}{A logical that specifies if plots should be created or not.
The default is \code{FALSE}.}

\item{col.weight}{The name of column that contains the weight of the fish
in kilograms. Units are important here because kilograms will be converted
to pounds to match other weight calculations.
The default is \code{weightkg}, which is created using \link{cleanPacFIN}.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is NULL.}
}
\value{
Additional columns are added to \code{Pdata}:
\itemize{
\item \code{Wt_Sampled_1}: the sum of sex-specific weights within the sample.
\item \code{Wt_Sampled_2}: the species-specific sample weight only provided by
California in cluster weight.
\item \code{LW_Calc_Wt}: individual weights predicted from the specified
length-weight relationships.
\item \code{Wt_Sampled_3}: The sum of empirical weights, for those fish within a
sample where this information is available, and weights calculated from the
length-weight relationship. This uses the empirical data if available and
fills in with the expected weight or mean-sample weight.
}
}
\description{
Calculate the denominator for the level-1 expansion, where the denominator
is the weight of all fish in the sample.
\code{EF1_Denominator} is not run by the user, it is a sub-function of
\link{getExpansion_1}.
}
\details{
The denominator of the level-1 expansion factor is the weight of the sampled
fish in the sample unit, where for fisheries data, this unit is most often
the trip level because information on individual hauls is not available.
For a survey, tow- or haul-level data are typically available.
The sum of the weight in the sample is calculated different for each state
based on the data that are available.

Oregon provides information on the weight of females and males in the sample
via the \code{WEIGHT_OF_FEMALES} and \code{WEIGHT_OF_MALES} columns. These are often
model-based weights, which would be the only way they can get weights when
the fish were not weighed themselves. The weight of unsexed fish is calculated
internally by the code and added to the female and male weight.
\strong{todo}: Let Oregon know that this calculation is being done and they may want
to provide WEIGHT_OF_UNKNOWN_LBS.

California sample weights were previously based on the column labeled
\code{SPECIES_WEIGHT}. Now, California data is parsed by PacFIN to furnish
species-specific cluster weights. Prior, cluster weights included the weight
of all species in the sample. Now, \code{CLUSTER_WEIGHT_LBS} is the weight in that
cluster for the species of interest. This was verified for Dover sole by
Kelli F. Johnson in February of 2021. Thus, the code uses \code{CLUSTER_WEIGHT_LBS}
rather than \code{SPECIES_WEIGHT} now.
\strong{todo}: determine if this should be the only cluster-specific value going forward?

Washington does not pretend to provide any information regarding total weight
of fish in the sample. Therefore, this value is calculated by summing the
empirical weight of fish. For fish that were not weighed, a weight is
calculated given the input parameters. It is up to the user to calculate
these if they want them based on more data than what is included in \code{Pdata}.
Finally, for fish that have no length and therefore no predicted weight,
the mean of fish weight in the sample is used to assign all fish a weight.

Another thing to note in the calculation of the denominator is that sample
weights are length- and age-specific values. You would not want to weight
an age sample based on how many lengths were measured. So, the weights of
fish that are not lengthed are backed out of the sample weight and the
weight of fish that are not aged are backed out of the sample weight to
create length- and age-specific sample weights. Only these specific sample
weights should be used going forward.
}
\seealso{
\link{EF1_Numerator}, \link{getExpansion_1}, \link{getExpansion_2}
}
\author{
Andi Stephens, Kelli F. Johnson, Chantel R. Wetzel
}
