% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GearTable}
\alias{GearTable}
\title{Table of gear types available in PacFIN data}
\format{
A data frame with 65 rows and 6 columns
\describe{
\item{TYPE}{A numeric}
\item{PACFIN_GEAR_CODE}{A three letter character string denoting the gear type}
\item{GRID}{A three letter character string denoting the gear type}
\item{GROUP}{A three letter character string that groups gear type into general groups}
\item{SHORT}{Character string describing the gear type}
\item{DESCRIPTION}{Character string describing the gear type}
}
}
\source{
\url{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/gr.txt}
}
\usage{
GearTable
}
\description{
Table of gear types available in PacFIN data
}
\keyword{datasets}
