% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latitude.R
\name{get_latitude}
\alias{get_latitude}
\title{Assign latitudes to records}
\usage{
get_latitude(Pdata, verbose = TRUE)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fisheries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
The data frame is returned with an additional column for latitude
}
\description{
Assign latitudes to records
}
\author{
Chantel Wetzel
}
