% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpansion_2.R
\name{getExpansion_2}
\alias{getExpansion_2}
\title{Second-stage expansion of composition samples up to catch level}
\usage{
getExpansion_2(
  Pdata,
  Catch,
  Units = "MT",
  Convert = lifecycle::deprecated(),
  maxExp = 0.95,
  stratification.cols,
  verbose = TRUE,
  savedir = NULL
)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fisheries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{Catch}{A data frame of catch data, in pounds or in metric tons.}

\item{Units}{The units of the \code{Catch} data frame, see
\code{measurements::conv_unit_options[["mass"]]} for options. Typical units
are metric tons (e.g., \code{"metric_ton"}) because that is the unit used
in Stock Synthesis, but expansions are done in pounds because fish weights
are in pounds. Thus, catches also need to be in pounds and will be
converted as such.}

\item{Convert}{A deprecated argument that is now set to \code{NULL}.
Previously, it was a logical that defined if the Catch should be converted from
metric tons to pounds, where \code{TRUE} is now the same as setting
\code{Units = "MT"} and \code{FALSE}, which was the default, would be
\code{Units = "LB"}. Normally, one would have their catch in metric tons,
i.e., \code{Convert = TRUE} or \code{Units = "MT"},
such that it can be used within Stock Synthesis.}

\item{maxExp}{The maximum expansion factor (either a number or a quantile)
for building expansions. Typically, the default is 0.95. Set \code{maxExp = Inf} to see largest values.}

\item{stratification.cols}{A vector of column names in \code{Pdata} that you want
to use as strata. These will match the way in which the catches are
transformed from long to wide prior to inputting them into this function.
If you leave this argument empty, then \code{Pdata} must already have a column
named \code{stratification}. The function will look in the column names of the
\code{Catch} data to determine the appropriate separator to use between columns
when pasting the words together, which is done using \link{apply} and \link{paste}.
Historically, it was mandatory to make this column yourself, but in 2021,
this input argument was added to reduce the number of extraneous calls that
were needed between functions. You can use as many levels of stratification
as you want except year because it is already included in the call to
\link[stats:aggregate]{stats::aggregate}.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}

\item{savedir}{where plots should be saved. The default
argument is NULL. If a directory is provided, then summary plots will be saved
to this location. If NULL, then no plots will be created.}
}
\value{
The input PacFIN dataset, with column \code{Expansion_Factor_2} appended.
}
\description{
The second-stage expansion calculates the expansion factor based on the
ratio of total catch within a stratification (e.g., year, gear group) to the
amount of that catch that was sampled.
}
\details{
Find the catch for each year and grouping in \code{Catch} and divide by the
pounds of fish that were collected for sampling for that same year and
grouping. Sampled biomass is stored in \code{All_Trips_Sampled_Lbs}, which is
the sum of Trip_Sampled_Lbs across sample numbers.
Catches were already stratified (i.e., summed by group placed in a column
for a given year or row). Catches are converted to pounds prior to dividing.
Thus, per-stratum Expansion_Factor_2 is the catch / sampled catch.
}
\section{Expansion}{

\itemize{
\item{Age data are expanded separately from lengths.}
\item{WA fish are generally only expanded using Expansion_Factor_2.}
\item{Other expansions are the product of
Expansion_Factor_1 * Expansion_Factor_2.
}
\item{For age-at-length comps, set Final_Expansion_Factor to 1 because
each fish represents only itself.}
}
}

\examples{
\dontrun{
data_exp2 <- getExpansion_2(
  Pdata = data_exp1,
  Catch = catch_formatted,
  Units = "MT",
  maxExp = 0.95
)
}

}
\seealso{
\itemize{
\item \link{getExpansion_1} should be ran first
}
}
\author{
Andi Stephens, Kelli F. Johnson, Chantel R. Wetzel
}
