% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGearGroup.R
\name{getGearGroup}
\alias{getGearGroup}
\title{Create column for gear called \code{geargroup} according to PacFIN gears}
\usage{
getGearGroup(Pdata, keep_gears = NULL, spp = NULL, verbose = TRUE)
}
\arguments{
\item{Pdata}{A data frame, typically one extracted from PacFIN but for this
function just one column is needed, \code{PACFIN_GEAR_CODE}.}

\item{keep_gears}{A character vector including the gear types you want
to label as unique fleets. Order matters and will define fleet numbering.
If the argument is missing, which is the default, then all found gear
groups are maintained and ordered alphabetically. For more details see
\code{\link[=getGearGroup]{getGearGroup()}}, which lists a link where you can find the available gear
groupings and how they link to \code{"PACFIN_GEAR_CODE"} within your data. The vector
supplied to this argument should consist of only options available in
\code{unique(GearTable[["GROUP"]])}.
Typical entries will include character values
such as \code{HKL}, \code{POT}, \code{TWL}, where the latter is short for all non-shrimp
trawls and \code{TWS} is shrimp trawls. Other gear identification codes and
their definitions include \code{DRG} which is dredge gear, \code{MSC} which is all
other miscellaneous gear such as diving or river trawls, \code{NET} which is all
non-trawl net gear, \code{NTW} which is non-trawl gear, and \code{TLS} which is
trolling gear. As a special case, \code{MID} is available for spiny dogfish to
extract mid-water trawl data as a separate fleet.}

\item{spp}{A character string giving the species name to ensure that the
methods are species specific. Leave \code{NULL} if generic methods work for
your species. Currently, sablefish is the only species with
species-specific code.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
A modified data frame where an additional column labeled \code{geargroup} is
added to \code{Pdata}. No original columns are modified in the process.
}
\description{
Create column for gear called \code{geargroup} according to PacFIN gears
}
\details{
Data from the PacFIN \href{http://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/gr.txt}{gear table}
is used to create a column in \code{Pdata} called \code{geargroup}, where
\code{Pdata[["PACFIN_GEAR_CODE"]]} is recoded to \code{geargroup} according to the gear group listed
in the table.
For some species the recoding is more intense. For example, sablefish
\code{GROUPS} are recoded into just three gears, hook and line (HKL), pot (POT),
and trawl (TWL). This is done internally to maintain consistency across
assessment authors. Other species can easily be added to this function. If a
species name is not passed then no additional recoding is performed compared
to the matching available in the table pulled from online.
}
\examples{
gears <- c("PRT", "FPT", "TWL", "MDP")
X <- getGearGroup(data.frame(PACFIN_GEAR_CODE = gears), verbose = TRUE)

}
\seealso{
\itemize{
\item \code{\link[=cleanPacFIN]{cleanPacFIN()}} uses this function
}
}
\author{
Andi Stephens, Kelli F. Johnson, Chantel R. Wetzel
}
