% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLength.R
\name{getLength}
\alias{getLength}
\title{Get length for PacFIN commercial samples}
\usage{
getLength(Pdata, keep, verbose = TRUE)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fisheries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{keep}{A vector of values that represent what you want to keep. Values
of \code{NA}, \code{""}, and numeric values are acceptable. Often, it is helpful to
run `unique(, useNA = "always") on the relevant data prior to running the
function to see what types are present.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
A vector of lengths in millimeters. Values of \code{NA} indicate that
the length should not be used in down-stream calculations.
}
\description{
Get length for PacFIN commercial samples in millimeters.
Lengths are converted for skates from type R and A to actual lengths.
Other, more standard lengths, are filtered for types in \code{keep}.
}
