% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getState.R
\name{getState}
\alias{getState}
\title{Create a state column based on input column specified in \code{source}}
\usage{
getState(Pdata, source = c("AGENCY_CODE"), verbose = TRUE)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fisheries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{source}{The column name where state information is located in
\code{Pdata}. See the function call for options, where only the first
value will be used.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
The input data frame is returned with an additional column,
\code{state}, which is filled with two-character values identifying the state or
a three-character value \code{UNK} for all rows that do not have an assigned state.
All rows are returned, but users should pay attention to the warning that is
returned for rows that have no state id.
}
\description{
Create a state column based on input column specified in \code{source}
}
\details{
Create a categorical column, \code{state}, based upon the \code{AGENCY_CODE}.
It is no longer advisable as of February 14, 2021 to create states based on
\code{PSMFC_CATCH_AREA_CODE} or \code{PSMFC_ARID} because areas are not mutually
exclusive to a state. Previous code set areas \code{1[a-z]} to Washington,
\code{2[a-z]} to Oregon, and \code{3[a-z]} to California.
The \href{https://pacfin.psmfc.org/wp-content/uploads/2019/03/PacFIN_Comprehensive_BDS_Commercial.pdf}{PacFIN documentation}
suggests that the following area codes can be assigned to the following states:
\itemize{
\item WA: 1C, 2A, 2B, 2C, 2E, 2F, 3A, 3B, 3C, 3N, 3S
\item OR: 1C, 2A, 2B, 2C, 2E, 2F, 3A, 3B, CS
\item CA: 1A, 1B, 1C
}

However, the look-up table for \href{https://apex.psmfc.org/pacfin/f?p=501:817:3123899543376:INITIAL}{PacFIN Catch Area Code}
indicates that they should be:
\itemize{
\item WA: 3A, 3B, 3C, 3S
\item OR: 2A, 2B, 2C, 2E, 2F, 3A
\item CA: 1A, 1B, 1C
}

If you see the need to use PSMFC_CATCH_AREA_CODE to set states
please contact the package maintainer.
}
\examples{
data <- data.frame(
  AGENCY_CODE = rep(c("W", "O", "C"), each = 2),
  info = 1:6
)
testthat::expect_true(
  all(getState(data)[["state"]] == rep(c("WA", "OR", "CA"), each = 2))
)

}
\seealso{
\link{cleanPacFIN} calls \code{getState}.
}
