% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacfin_ports_withlatlong}
\alias{pacfin_ports_withlatlong}
\title{Data frame of pacfin port names and latitudes}
\format{
A data frame with 511 rows and 6 columns
\describe{
\item{name}{U.S. West Coast port name.}
\item{pcid}{Port code used in PacFIN}
\item{agencydesc}{State agency port name.}
\item{agid}{State agency identifier where W = Washington, O = Oregon, and C = California.}
\item{longitude}{Longitude of the port location with minutes and seconds in radians.s}
\item{latitude}{Latitude of the port location with minutes and seconds in radians.}
}
}
\usage{
pacfin_ports_withlatlong
}
\description{
Data frame of pacfin port names and latitudes
}
\keyword{datasets}
