% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCleaned.R
\name{plotCleaned}
\alias{plotCleaned}
\title{Plot cleaned fishery age- or length-composition data.}
\usage{
plotCleaned(Pdata, savedir)
}
\arguments{
\item{Pdata}{A data object created using \code{\link[=cleanPacFIN]{cleanPacFIN()}}].}

\item{savedir}{Directory where output will be saved. The directory where the file should be saved.}
}
\value{
The following figures are saved to \code{savedir}:
\itemize{
\item PacFIN_comp_Nbystate.png
\item PacFIN_comp_distributions.png
\item PacFIN_comp_NbyGRID.png
\item PacFIN_comp_depth.png
\item PacFIN_comp_geargroup.png
\item PacFIN_comp_INPFC.png
\item PacFIN_comp_lengthvage.png
}
}
\description{
Save several figures to your computer that summarize the composition
samples that are available for the given species. None of the figures
are meant to be included in the stock assessment document, but rather,
are for exploring the data.
}
\seealso{
This function is called by \link{cleanPacFIN} and heavily
relies on \link{getGearGroup} to create gear categories.
}
\author{
Andi Stephens, Kelli F. Johnson
}
