% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWL.R
\name{plotWL}
\alias{plotWL}
\title{Plot of the weight--length relationship by sex}
\usage{
plotWL(
  length,
  sex,
  weight,
  weight.calc,
  xlab = "Length (cm)",
  ylab = "Weight (kg)"
)
}
\arguments{
\item{length}{A vector of fish lengths in mm.}

\item{sex}{A vector of sexes for each entry in \code{length}.
Entries must be one of the following: 'U', 'F', 'M', or 'H'.}

\item{weight}{A vector of empirical weights in kilograms.}

\item{weight.calc}{A vector of estimated weights output from \link{getweight}.}

\item{xlab, ylab}{A character value specifying the label for an axes of
a figure. This is typically passed to \link[ggplot2:xlab]{ggplot2::xlab} or \link[ggplot2:ylab]{ggplot2::ylab}
but it can also be passed to the \code{xlab} or \code{ylab} arguments of \link{plot}.
No additional formatting will take place.}
}
\description{
Plot the weight--length relationship using \link[ggplot2:ggplot]{ggplot2::ggplot}.
Sex-specific relationships are plotted if available using shades of gray.
Note that \code{sex} can be missing and a single relationship for all data
will be plotted.
}
\seealso{
\link{getWLpars} to calculate parameters of the weight--length relationship;
\link{getweight} to predict weights from empirical lengths
}
\author{
Kelli F. Johnson
}
