% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PortTable}
\alias{PortTable}
\title{Table of ports available in PacFIN data}
\format{
A data frame with 567 rows and 4 columns
\describe{
\item{PACFIN_PORT_CODE}{A character string of port codes}
\item{AGENCY_CODE}{Character denoting the collection agency}
\item{AGENCY_PORT_CODE}{A string denoting the sampling agency port code}
\item{DESCRIPTION}{A character string denoting the full port name}
}
}
\source{
\url{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/agency_ports_pcid.txt}
}
\usage{
PortTable
}
\description{
Table of ports available in PacFIN data
}
\keyword{datasets}
