% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INPFCTable}
\alias{INPFCTable}
\title{Table of INPFC areas available in PacFIN data}
\format{
A data frame with 69 rows and 7 columns
\describe{
\item{INPFC_AREA_TYPE_CODE}{A character string}
\item{ARID}{A character string}
\item{INPFC}{A three letter character string denoting the INPFC area}
\item{TYPE}{A character string denoting the INPFC type}
\item{COUNCIL}{A character string denoting the Fishery Managment Council}
\item{SHORT}{Character string describing the area}
\item{DESCRIPTION}{Character string describing the area}
}
}
\source{
\url{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/ar_tree.txt}
}
\usage{
INPFCTable
}
\description{
Table of INPFC areas available in PacFIN data
}
\keyword{datasets}
